+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 5.6.14-SNAPSHOT / WIP
 ==================
+* [OLMIS-8053](https://openlmis.atlassian.net/browse/OLMIS-8053): Enable the integration of country-level population data into the system
 
 5.6.13 / 2024-10-31
 ==================